Druntime: Select newer NetBSD time bindings (#22720)
* Correct NetBSD mangle for core.sys.posix.stdc.time

* Correct NetBSD mangle for core.sys.posix.time

* Correct NetBSD bindings for core.sys.posix.sys.select

* Correct NetBSD bindings for core.sys.posix.sys.statvfs

* Correct NetBSD bindings for core.sys.posix.sys.time

* time_t is 64-bit on with newer signatures

* clock_t is unsigned int, not c_long

* suseconds_t is int, not c_long

+++ druntime/src/core/sys/posix/sys/types.d
@@ -251,7 +251,7 @@ else version (NetBSD)
     alias pid_t = int;
     //size_t (defined in core.stdc.stddef)
     alias ssize_t = c_long;
-    alias time_t = c_long;
+    alias time_t = long; // _BSD_TIME_T_ mapped to __int64_t
     alias uid_t = uint;
 }
 else version (OpenBSD)
@@ -388,10 +388,10 @@ else version (NetBSD)
 {
     alias fsblkcnt_t = ulong;
     alias fsfilcnt_t = ulong;
-    alias clock_t = c_long;
+    alias clock_t = uint; // unsigned int
     alias id_t = long;
     alias key_t = c_long;
-    alias suseconds_t = c_long;
+    alias suseconds_t = int; // int
     alias useconds_t = uint;
 }
 else version (OpenBSD)
